+++ cmd/serf/command/force_leave.go
@@ -17,8 +17,11 @@ type ForceLeaveCommand struct {
 var _ cli.Command = &ForceLeaveCommand{}
 
 func (c *ForceLeaveCommand) Run(args []string) int {
+	var prune bool
+
 	cmdFlags := flag.NewFlagSet("join", flag.ContinueOnError)
 	cmdFlags.Usage = func() { c.Ui.Output(c.Help()) }
+	cmdFlags.BoolVar(&prune, "prune", false, "Remove agent forcibly from list of members")
 	rpcAddr := RPCAddrFlag(cmdFlags)
 	rpcAuth := RPCAuthFlag(cmdFlags)
 	if err := cmdFlags.Parse(args); err != nil {
@@ -39,6 +42,16 @@ func (c *ForceLeaveCommand) Run(args []string) int {
 		return 1
 	}
 	defer client.Close()
+
+	if prune {
+		err = client.ForceLeavePrune(nodes[0])
+		if err != nil {
+			c.Ui.Error(fmt.Sprintf("Error force leaving: %s", err))
+			return 1
+		}
+		return 0
+
+	}
 
 	err = client.ForceLeave(nodes[0])
 	if err != nil {
@@ -68,6 +81,7 @@ Options:
 
   -rpc-addr=127.0.0.1:7373  RPC address of the Serf agent.
   -rpc-auth=""              RPC auth token of the Serf agent.
+  -prune                    Remove agent forcibly from list of members
 `
 	return strings.TrimSpace(helpText)
 }